fix(templates) always set SSL verify depth
The existing template will only set `lua_ssl_verify_depth` if
`lua_ssl_trusted_certificate` is set also. It should be possible to set
verification depth regardless of whether custom CAs are in use. Note
that this change will always include this directive, even if using the
default (1). This should be the case, as Kong's default depth may
diverge from OpenResty's in the future.